requirements.txt
selenium>=4.0.0
pytest
pytest-html
python-dotenv